conver_bot/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
frozenlist==1.5.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-speech==2.30.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
mariadb==1.1.11
multidict==6.1.0
mysql-connector-python==9.1.0
openai==1.59.7
packaging==24.2
pillow==11.1.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pygame==2.6.1
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
svgwrite==1.4.3
tqdm==4.67.1
Tree==0.2.4
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
